+++ lib/src/domain/entities/prizes/prizes.dart
@@ -1,3 +1,4 @@
+export 'coupon.dart';
 export 'reward.dart';
 export 'reward_check.dart';
 export 'reward_redeem.dart';